Refactor controller to use events instead of polling for service status
