+++ dask_kubernetes/operator/controller/controller.py
@@ -1,4 +1,3 @@
-import asyncio
 from datetime import datetime
 
 import aiohttp
@@ -72,6 +71,7 @@ def build_scheduler_service_spec(cluster_name, spec, annotations):
             "name": f"{cluster_name}-scheduler",
             "labels": {
                 "dask.org/cluster-name": cluster_name,
+                "dask.org/component": "scheduler",
             },
             "annotations": annotations,
         },
@@ -182,26 +182,6 @@ def build_cluster_spec(name, worker_spec, scheduler_spec):
     }
 
 
-async def wait_for_service(api, service_name, namespace):
-    """Block until service is available."""
-    while True:
-        try:
-            service = await api.read_namespaced_service(service_name, namespace)
-
-            # If the service is of type LoadBalancer, also wait until it's ready.
-            if (
-                service.spec.type == "LoadBalancer"
-                and len(service.status.load_balancer.ingress or []) == 0
-            ):
-                pass
-            else:
-                break
-        except Exception:
-            pass
-        finally:
-            await asyncio.sleep(0.1)
-
-
 @kopf.on.startup()
 async def startup(**kwargs):
     await ClusterAuth.load_first()
@@ -219,7 +199,7 @@ async def daskcluster_create(name, namespace, logger, patch, **kwargs):
 
 @kopf.on.field("daskcluster", field="status.phase", new="Created")
 async def daskcluster_create_components(spec, name, namespace, logger, patch, **kwargs):
-    """When the DaskCluster status.phase goes into Pending create the cluster components."""
+    """When the DaskCluster status.phase goes into Created create the cluster components."""
     logger.info("Creating Dask cluster components.")
     async with kubernetes.client.api_client.ApiClient() as api_client:
         api = kubernetes.client.CoreV1Api(api_client)
@@ -234,7 +214,6 @@ async def daskcluster_create_components(spec, name, namespace, logger, patch, **
             namespace=namespace,
             body=data,
         )
-        # await wait_for_scheduler(name, namespace)
         logger.info(f"Scheduler pod {data['metadata']['name']} created in {namespace}.")
 
         # TODO Check for existing scheduler service
@@ -260,8 +239,35 @@ async def daskcluster_create_components(spec, name, namespace, logger, patch, **
             body=data,
         )
         logger.info(f"Worker group {data['metadata']['name']} created in {namespace}.")
-    # TODO Set to "Pending" here and track scheduler readiness before finally setting to "Running"
-    patch.status["phase"] = "Running"
+    patch.status["phase"] = "Pending"
+
+
+@kopf.on.field("service", field="status", labels={"dask.org/component": "scheduler"})
+async def handle_scheduler_service_status(spec, meta, status, namespace, **kwargs):
+    # If the Service is a LoadBalancer with no ingress endpoints mark the cluster as Pending
+    if spec["type"] == "LoadBalancer" and not len(
+        status.get("load_balancer", {}).get("ingress", [])
+    ):
+        phase = "Pending"
+    # Otherwise mark it as Running
+    else:
+        phase = "Running"
+
+    async with kubernetes.client.api_client.ApiClient() as api_client:
+        customobjectsapi = kubernetes.client.CustomObjectsApi(api_client)
+        api_client.set_default_header("content-type", "application/merge-patch+json")
+        await customobjectsapi.patch_namespaced_custom_object_status(
+            group="kubernetes.dask.org",
+            version="v1",
+            plural="daskclusters",
+            namespace=namespace,
+            name=meta["labels"]["dask.org/cluster-name"],
+            body={
+                "status": {
+                    "phase": phase,
+                }
+            },
+        )
 
 
 @kopf.on.create("daskworkergroup")
