+++ dask_kubernetes/operator/kubecluster/kubecluster.py
@@ -26,9 +26,6 @@ from distributed.utils import (
 )
 
 from dask_kubernetes.common.auth import ClusterAuth
-from dask_kubernetes.operator.controller import (
-    wait_for_service,
-)
 
 from dask_kubernetes.common.networking import (
     get_scheduler_address,
@@ -841,6 +838,26 @@ def make_scheduler_spec(
             ],
         },
     }
+
+
+async def wait_for_service(api, service_name, namespace):
+    """Block until service is available."""
+    while True:
+        try:
+            service = await api.read_namespaced_service(service_name, namespace)
+
+            # If the service is of type LoadBalancer, also wait until it's ready.
+            if (
+                service.spec.type == "LoadBalancer"
+                and len(service.status.load_balancer.ingress or []) == 0
+            ):
+                pass
+            else:
+                break
+        except Exception:
+            pass
+        finally:
+            await asyncio.sleep(0.1)
 
 
 @atexit.register
